--- a/lib/benchmark/ips.rb
+++ b/lib/benchmark/ips.rb
@@ -1,6 +1,5 @@
 # encoding: utf-8
 require 'benchmark/timing'
-require 'benchmark/compare'
 require 'benchmark/ips/stats/stats_metric'
 require 'benchmark/ips/stats/sd'
 require 'benchmark/ips/stats/bootstrap'