--- a/lib/benchmark/ips/json_report.rb
+++ b/lib/benchmark/ips/json_report.rb
@@ -1,0 +1,19 @@
+module Benchmark
+  module IPS
+    class JsonReport
+      def initialize(json_path)
+        @json_path = json_path
+      end
+
+      def run(job)
+      end
+
+      def post_run(report)
+        File.open @json_path , "w" do |f|
+          require "json"
+          f.write JSON.pretty_generate(report.data)
+        end
+      end
+    end
+  end
+end