--- a/lib/benchmark/ips/report.rb
+++ b/lib/benchmark/ips/report.rb
@@ -174,20 +174,6 @@ module Benchmark
           }
         end
       end
-
-      # Run comparison of entries.
-      def run_comparison
-        Benchmark.compare(*@entries)
-      end
-
-      # Generate json from Report#data to given path.
-      # @param path [String] path to generate json.
-      def generate_json(path)
-        File.open path, "w" do |f|
-          require "json"
-          f.write JSON.pretty_generate(data)
-        end
-      end
     end
   end
 end