Hi @evanphx 

I find I often want to extend `benchmark-ips` to add number of objects allocated or number of queries performed and others.

There are a few memory alternatives like https://github.com/michaelherold/benchmark-memory (there was another, can't find it), but to be honest, it would be nice if there were just one base for this mostly common code. I was hoping to keep `benchmark-ips` as the base glue that would be able to run multiple types of benchmarks. (probably not your main goal)

I think the first 2 commits are no-brainers.
Possibly the 3rd is as well.
If you prefer, I can pull some of those out into a separate PR and go from there.