@evanphx 2018-08-28
This is great! Really no qualms about adding it. Do you have an memory one maybe we could add to the release too?