@kbrock 2018-09-01
I do wondering about the overall concept of hold.

You can run 2 kinds of benchmarks (probably more but...)

A. 1 file with a bunch of implementations of the same method. e.g.: [fast-ruby](https://github.com/JuanitoFatas/fast-ruby)
B. 1 file with a bunch of implementations of different methods. e.g.: [ruby bench](https://github.com/ruby-bench/ruby-bench)

It seems that `hold` is for case B. But it becomes very manual. Improving a method and then running with 2 implementations seems to add quite a few steps. May be nice to run each version of the library from a CI.

Does it makes more sense to just use an api (i.e.: benchmark.fyi) for running across multiple implementations? This is essentially what the `hold.csv` file is. Right? common data across multiple runs. it can be a save file or a process that has an API.

We would need more blob/json based stats to store memory, queries, timings.
We would need meta-data on the implementation that was run for this particular test. e.g.: ruby version, library version, time run.
But besides that, it is very similar to benchmark.fyi

Think I'm getting ahead of myself though.
And maybe I'm just reinventing ruby-bench.