@kbrock 2018-09-05
@evanphx Hmm. This is a great conversation. But I'm sorry I conflated these 2 conversation here. I'll open another issue or pr with some thoughts.

@michaelherold I would like to define the examples once, and then use them across both / other test types. I had thought this could be extensible through the `Suite` interface. But realized the `Suite` is more focused on what happens between tests, and this is more focused on what happens in an individual test.