@michaelherold 2018-09-06
@evanphx It does that job very well -- that's what I use it for, in fact. 😄