@kbrock 2018-10-18
ok, pushed up some changes.
I'm glad you asked me to extend this, because I found a few issues in the implementation.

1. reworked more of this than I had expected.
2. Not sure if the helper interface should be `run(full_report)` or `run(job)`.
3. Not sure if you want me to add `Benchmark::IPS::Objects`. Seems work from @michaelherold is more thorough. Hope it will be a simple example if nothing else.

```ruby
require "benchmark/ips"

# comparison helper
# current interface passes a report object to run (since all seem to use full_report.entries or full_report.data)
# we need job.list, so we pass it as part of the initializer (could change the interface to run)
class CompareObjects
  def initialize(job)
    @job = job
  end

  # Compare between reports, prints out facts of each report:
  # objects allocated
  # @param report [Benchmark::IPS::Report] Report with entries to compare.
  def run(_report)
    entries = @job.list
    return if entries.size < 2

    objects_key = RUBY_VERSION < '2.2' ? :total_allocated_object : :total_allocated_objects

    allocations = {}

    entries.each do |entry|
      before = GC.stat[objects_key]
      entry.call_times(1)
      after = GC.stat[objects_key]
      allocations[entry] = after - before
    end

    sorted = entries.sort_by{ |e| allocations[e] }.reverse

    best = sorted.shift

    $stdout.puts "\nComparison:"
    $stdout.printf "%20s: %10d objects\n", best.label, allocations[best]

    sorted.each do |report|
      name = report.label.to_s
      $stdout.printf "%20s: %10d\n", name, allocations[report]
      $stdout.puts
    end

    $stdout.puts
  end
end

# under test

class Def
  def initialized
    @defined = true
  end

  # yes, I'm intentionally setting a different variable
  # want it to remain undefined

  # this is the winner (they are all very close)
  def undefqc # question colon 
    defined?(@undefined) ? @undefined : @defined = true
  end

  def defqc
    defined?(@defined) ? @defined : @defined = true
  end
end

d = Def.new

# benchmark

Benchmark.ips(warmup: 1, time: 1) do |x|
  x.report("undef ?:") { d.undefqc }
  x.report("def ?:") { d.defqc }
  #x.compare!
  x.add_helper(:objects, CompareObjects.new(x))
end
```

```
ruby -Ilib cast_def_microbenchmark.rb 
Warming up --------------------------------------
warmup: Benchmark::IPS::Job::Entry (true)
            undef ?:   267.706k i/100ms
warmup: Benchmark::IPS::Job::Entry (true)
              def ?:   273.221k i/100ms
Calculating -------------------------------------
            undef ?:      8.716M (± 2.9%) i/s -      8.834M in   1.014423s
              def ?:      8.878M (± 3.4%) i/s -      9.016M in   1.016769s

Comparison:
              def ?::          1 objects
            undef ?::          1
```