@kbrock 2018-10-19
ok - I'm going back in.

I'll put this into a memory report - hoping michael can tear my implementation out and make it as great as his gem

I feel the data should go into the `Report::Entry` rather than just printing it out