@kbrock 2018-10-19
```ruby
require "benchmark/ips"
require "active_record"

ActiveRecord::Base.establish_connection(ENV.fetch('DATABASE_URL') { "postgres://localhost/user_benchmark" })
ActiveRecord::Migration.verbose = false

ActiveRecord::Schema.define do
  create_table :users, force: true do |t|
    t.string :name
  end
  #add_index :users, :name
end
class User < ActiveRecord::Base ;end

if User.count == 0
  puts "Creating 100 users"
  100.times { |i| User.create name: "user #{i}" }
end

Benchmark.ips do |x|
  x.report("User.all.first") { User.all.first }
  x.report("User.all.to_a.first") { User.all.to_a.first }

  x.compare!
  x.queries!
  x.objects!
end
```

```
Creating 100 users
Warming up --------------------------------------
      User.all.first   287.000  i/100ms
 User.all.to_a.first    81.000  i/100ms
Calculating -------------------------------------
      User.all.first      2.962k (±14.5%) i/s -     14.637k in   5.081636s
 User.all.to_a.first    955.136  (± 7.2%) i/s -      4.779k in   5.034635s
Queries:
      User.all.first:   1 queries     1 objects
 User.all.to_a.first:   1 queries   100 objects

Comparison:
      User.all.first:     2962.3 i/s
 User.all.to_a.first:      955.1 i/s - 3.10x  slower

Queries Comparison:
 User.all.to_a.first:   1 queries   100 objects
      User.all.first:   1 queries     1 objects


Comparison:
 User.all.to_a.first:       1184 objects
      User.all.first:        148 objects
```